backend/requirements.txt
# ============================================
# Platforma - Python Dependencies
# ============================================

# Django
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=23.5

# Database
psycopg2-binary>=2.9

# Environment
python-decouple>=3.8

# Media
Pillow>=10.2

# Server
gunicorn>=21.2

# Utils
python-dateutil>=2.8
